+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - **Package Handler** Refactor: Consolidate two functions regarding getting source url and source tag into one
 - **Entire Script** Refactor: Standardize logging output across all files
 - **Package Handler** Refactor: Simplified package source url old and new
+- **Package Handler** Added the ability to get the changelog for git.kernel.org packages
 
 ### Bug fixes